Recovery level

+++ Assets/Scripts/Data/LevelData.cs
@@ -9,7 +9,7 @@ namespace Assets.Scripts.Data
     {
         public Vector2IntData gridSize;
         public List<ArrowData> arrows;
-        public int duration; // Optional: duration in seconds for time-based levels (0 = no time limit)
+        public float duration; // Optional: duration in seconds for time-based levels (0 = no time limit)
     }
 
     [Serializable]
